# Conflicts:
#	src/main/java/com/example/repair/config/myMetaObjectHandler.java
#	src/main/java/com/example/repair/controller/administer/AdministerController.java
#	src/main/java/com/example/repair/controller/student/studentController.java

+++ src/main/java/com/example/repair/entity/AdministratorAccount.java
@@ -30,7 +30,7 @@ public class AdministratorAccount implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "job_number", type = IdType.INPUT)
-    private Integer jobNumber;
+    private Long jobNumber;
 
     private String name;
